汎用的な関数calculateTimeAgoをlib/utilに移動。
